refactor(ServerProperties): 공통으로 사용중인 서버의 환경변수를 Common 패키지 하부로 이동
