+++ src/main/java/com/kustacks/kuring/auth/secured/SecuredAnnotationChecker.java
@@ -12,7 +12,6 @@ import org.springframework.stereotype.Component;
 import java.lang.reflect.Method;
 import java.util.Arrays;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Aspect
 @Component
